+++ libs/ui/package.json
@@ -88,6 +88,7 @@
     "eslint-plugin-react": "^7.18.3",
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-vx": "workspace:*",
+    "fast-check": "2.18.0",
     "is-ci-cli": "^2.2.0",
     "jest": "^27.3.1",
     "jest-watch-typeahead": "^0.6.4",
